fix(trabalhos/t1): cast count column to integer

+++ trabalhos/t1/src/analysis.py
@@ -1,8 +1,5 @@
 import json
 import pandas as pd
-import numpy as np
-import matplotlib.pyplot as plt
-import networkx as nx
 
 from mlxtend.preprocessing import TransactionEncoder
 from mlxtend.frequent_patterns import apriori, association_rules
@@ -32,7 +29,7 @@ def create_rules(df, min_support=0.2, metric='confidence', min_threshold=0.0):
     frequent_itemsets = apriori(df, min_support=min_support, use_colnames=True)
     rules = association_rules(frequent_itemsets, metric=metric, min_threshold=min_threshold)
     
-    rules['count'] = round(rules['support'] * len(df.index))
+    rules['count'] = round(rules['support'] * len(df.index)).astype(int)
 
     return rules.sort_values(by=metric, ascending=False)
